Consolidated tests
