+++ tests/Kentico.Xperience.AzureSearch.Tests/Builder/TestSearchModelWithAdditionalPropertiesStrategy.cs
@@ -1,0 +1,7 @@
+using Kentico.Xperience.AzureSearch.Indexing;
+
+namespace Kentico.Xperience.AzureSearch.Tests;
+
+internal class TestSearchModelWithAdditionalPropertiesStrategy : BaseAzureSearchIndexingStrategy<TestSearchModelWithAdditionalProperties>
+{
+}
